Add ParseArgs command line flag
Adding a ParseArgs command line flag to gearcmd to support sending json
as payloads.

For example, before this change the json payload {"key":"value"} would
get parsed by bash into {key:value} (ie without the quotes) which is
invalid json. We decided the easiest way to address this was to add a
flag that allowed the user to bypass any parsing and send the raw
payload to the worker.